+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## [2.0.3](https://github.com/bitrix24/b24ui/compare/v2.0.2...v2.0.3) (2025-10-24)
 
-* feat(InputNumber): handle `increment` / `decrement` as booleans
-* fix(Error): render as `div` instead of `main`
-* feat(SidebarLayout/DashboardGroup): improve sidebarLoading hook
+### Features
+* **InputNumber:** handle `increment` / `decrement` as booleans
+* **SidebarLayout/DashboardGroup:** improve sidebarLoading hook
+
+### Bug Fixes
+* **Error:** render as `div` instead of `main`
 
 ## [2.0.2](https://github.com/bitrix24/b24ui/compare/v2.0.1...v2.0.2) (2025-10-23)